+++ src/events/ResponseEvent.php
@@ -1,0 +1,17 @@
+<?php
+namespace wheelform\events;
+
+use yii\base\Event;
+
+class ResponseEvent extends Event
+{
+    public $headers;
+
+    public $message;
+
+    public $success;
+
+    public $data;
+
+    public $errors;
+}