+++ src/services/MetaTagsService.php
@@ -1,0 +1,13 @@
+<?php
+namespace wheelform\services;
+
+use craft\helpers\Template;
+use yii\helpers\Html;
+
+class MetaTagsService extends BaseService
+{
+    public function __invoke()
+    {
+        return Template::raw(Html::csrfMetaTags());
+    }
+}