+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 2.4.3 - 2020-06-29
 ### Added
 - Ability to enable/disable "required" attribute for most fields.
+### Fixed
+- Fixed bug related to console purge not running correctly
 
 ## 2.4.2 - 2020-02-28
 ### Added